Making Expense App Production ready Task 13

+++ controllers/expense.js
@@ -1,7 +1,8 @@
 const Expense = require('../models/expense');
 const User = require('../models/user');
+const sequelize = require('../util/database');
 exports.addExpense = async (req, res, next)=> {
-
+    const t = await sequelize.transaction();
    try{   
 
    const amount = req.body.amount;
@@ -19,28 +20,32 @@ exports.addExpense = async (req, res, next)=> {
    if(amount == undefined || amount.length === 0) {
       return res.status(400).json({success: false, message: 'Parameters missing'})
   }
-   await Expense.create({ amount, description, category, userId: req.user.id})
-   .then(expense => {
+   const expense = await Expense.create({ amount, description, category, userId: req.user.id},{transaction: t})
+   //.then(expense => {
       console.log("amount",amount);
       console.log("req.user.totalExp",req.user.totalExpenses);
       const totalExpense = Number(req.user.totalExpenses) + Number(amount)
       console.log("totalExpense",totalExpense);
-      User.update({
+     await User.update({
           totalExpenses: totalExpense
       },{
-          where: {id: req.user.id} 
-      }).then(async() => {
-          res.status(200).json({newExpense: expense})
+          where: {id: req.user.id} , transaction: t
       })
-      .catch(async(err) => {
-          return res.status(500).json({success: false, error: err})
-      })
-    }).catch(async(err) => {
-       return res.status(500).json({success: false, error: err})
-    })
+    //   .then(async() => {
+    //       res.status(200).json({newExpense: expense})
+    //   })
+    //   .catch(async(err) => {
+    //       return res.status(500).json({success: false, error: err})
+    //   })
+    // }).catch(async(err) => {
+    //    return res.status(500).json({success: false, error: err})
+    // })
+        await t.commit();
+        res.status(200).json({newExpense: expense})
     } catch(err) {
-      console.log(`posting data is not working`);
-      res.status(500).json(err);
+        await t.rollback();
+        console.log(`posting data is not working`);
+        res.status(500).json(err);
     }
 
 }
@@ -57,6 +62,7 @@ exports.getExpense = async (req, res, next) => {
 }
 
 exports.deleteExpense = async (req, res) => {
+    const t = await sequelize.transaction();
     const eId = req.params.id;
     console.log(req.params.id);
     try{
@@ -64,10 +70,35 @@ exports.deleteExpense = async (req, res) => {
        console.log('ID is missing');
       return res.status(400).json({err: 'ID is missing'})
     }
-    await Expense.destroy({where: {id: eId}});
-    res.sendStatus(200);
-    } catch(err){
-       console.log(err);
-       res.status(500).json(err)
+    // await Expense.destroy({where: {id: eId}});
+    // res.sendStatus(200);
+    // } catch(err){
+    //    console.log(err);
+    //    res.status(500).json(err)
+    // }
+    const expense = await Expense.findOne({where: {
+        id: eId,
+        userId: req.user.id 
+       }
+      },{transaction:t});
+      const totalExpenses = await Expense.sum('amount', {
+        where: {userId: req.user.id}
+    })
+    const updatedTotalExpenses = totalExpenses - expense.amount
+    const noOfRows = await Expense.destroy({where: {id: eId, userId: req.user.id}});
+    await User.update({
+      totalExpenses: updatedTotalExpenses
+    },{
+      where: {id: req.user.id},
+    })
+    if(noOfRows === 0) {
+      return res.status(404).json({message: `Expense doesn't belongs to user`})
     }
+    await t.commit();
+    res.sendStatus(200);
+      } catch(err){
+        await t.rollback();
+        console.log(err);
+        res.status(500).json(err)
+      }
 }
